requirements.txt
pandas 
numpy
sqlalchemy
python-dotenv
mysql-connector-python
pymysql
scikit-learn
##-e .